python_playground/pyproject.toml
[tool.poetry]
name = "python_playground"
version = "0.1.0"
description = ""
authors = ["david <david@cdataiapp.com>"]

[tool.poetry.dependencies]
python = "^3.8"
protobuf = "^4.21.12"
numpy = "^1.24.1"
sympy = "^1.11.1"
tqdm = "^4.64.1"
Pillow = "^9.4.0"
pandas = "^1.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
